feat: 로또 다시시작 기능 구현하기
