+++ index.html
@@ -111,7 +111,7 @@
           </div>
           <p class="text-center font-bold">당신의 총 수익률은 <span class="earning-rate">오오오%</span>입니다.</p>
           <div class="d-flex justify-center mt-5">
-            <button type="button" class="btn btn-cyan">다시 시작하기</button>
+            <button type="button" class="btn btn-cyan btn-restart">다시 시작하기</button>
           </div>
         </div>
       </div>
