+++ src/js/util/Util.js
@@ -1,0 +1,1 @@
+export const getNumbers = (arr = []) => arr.map(Number);
